requirements.txt
jsonschema
pillow